margins and dpi are double values
git-svn-id: svn://svn.code.sf.net/p/iup/iup/trunk@4323 2f08bc20-984a-4df8-b195-9b5d5d477537

+++ iup/srcscintilla/iupsci.h
@@ -65,7 +65,7 @@ int iupdrvScintillaGetBorder(void);
 int iupdrvScintillaPrintAttrib(Ihandle* ih, const char* value);
 enum { PRINTUNITS_PIXELS, PRINTUNITS_INCH, PRINTUNITS_CM };
 int iupSciGetPrintMarginUnits(Ihandle* ih);
-int iupSciGetPrintMargin(Ihandle* ih, const char* margin_attrib, int units, int dpi);
+int iupSciGetPrintMargin(Ihandle* ih, const char* margin_attrib, int units, double dpi);
 
 
 #ifdef __cplusplus
